chore: move EEbus to global config
