+++ cmd/configure/eebus.go
@@ -3,8 +3,8 @@ package configure
 import (
 	"fmt"
 
-	"github.com/evcc-io/evcc/charger/eebus"
 	"github.com/evcc-io/evcc/cmd/shutdown"
+	"github.com/evcc-io/evcc/server/eebus"
 	"github.com/evcc-io/evcc/util"
 )
